fixed inscription

+++ siteConfinement/ConfinAide/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class Client(models.Model):
-    id_client = models.BigIntegerField(primary_key=True)
+    id_client = models.AutoField(primary_key=True, )
     password = models.CharField(max_length = 20)
     mail = models.CharField(max_length = 50)
     nom = models.CharField(max_length = 20, null = True)
